Allow makeDocs to understand quoted paths in doxygen config files
Previously the presence of a double quote was taken as a sign that
the string should be passed through to the output and not be
processed. Now items known to represent file paths are now understood
to be able to contain double quotes when spaces may be present.

Output paths are also now quoted if they contain spaces.